+++ lib/services/bounty_service.dart
@@ -14,12 +14,12 @@ class BountyService {
   final ClientService _clientService;
   final GithubService _githubService;
 
-  Future<Bounty> getBounty(int bountyId) async {
+  Future<Bounty> getBounty(BigInt bountyId) async {
     final info = await _clientService.getBounty(bountyId);
     final issue = await _githubService.getIssue(
       info.repoOwner,
       info.repoName,
-      info.issueNumber,
+      info.issueNumber.toInt(),
     );
     if (issue == null) {
       return null;
@@ -27,12 +27,12 @@ class BountyService {
     return Bounty(info: info, issue: issue);
   }
 
-  Future<BountySubmission> getSubmission(int submissionId) async {
+  Future<BountySubmission> getSubmission(BigInt submissionId) async {
     final info = await _clientService.getSubmission(submissionId);
     final issue = await _githubService.getIssue(
       info.repoOwner,
       info.repoName,
-      info.issueNumber,
+      info.issueNumber.toInt(),
     );
     if (issue == null) {
       return null;
@@ -40,14 +40,14 @@ class BountyService {
     return BountySubmission(info: info, issue: issue);
   }
 
-  Future<List<Bounty>> listOpenBounties(int min) async {
+  Future<List<Bounty>> listOpenBounties(BigInt min) async {
     final list = await _clientService.listOpenBounties(min);
     final result = <Bounty>[];
     for (final info in list) {
       final issue = await _githubService.getIssue(
         info.repoOwner,
         info.repoName,
-        info.issueNumber,
+        info.issueNumber.toInt(),
       );
       if (issue == null) {
         continue;
@@ -58,7 +58,7 @@ class BountyService {
   }
 
   Future<List<BountySubmission>> listBountySubmissions(
-    int bountyId,
+    BigInt bountyId,
   ) async {
     final list = await _clientService.listBountySubmissions(bountyId);
     final result = <BountySubmission>[];
@@ -66,7 +66,7 @@ class BountyService {
       final issue = await _githubService.getIssue(
         info.repoOwner,
         info.repoName,
-        info.issueNumber,
+        info.issueNumber.toInt(),
       );
       if (issue == null) {
         return null;
@@ -76,11 +76,11 @@ class BountyService {
     return result;
   }
 
-  Future<int> postBounty(
+  Future<BigInt> postBounty(
     String repoOwner,
     String repoName,
-    int issueNumber,
-    int amount,
+    BigInt issueNumber,
+    BigInt amount,
   ) {
     return _clientService.postBounty(
       repoOwner,
@@ -90,12 +90,12 @@ class BountyService {
     );
   }
 
-  Future<int> submitForBounty(
-    int bountyId,
+  Future<BigInt> submitForBounty(
+    BigInt bountyId,
     String repoOwner,
     String repoName,
-    int issueNumber,
-    int amount,
+    BigInt issueNumber,
+    BigInt amount,
   ) {
     return _clientService.submitForBounty(
       bountyId,
@@ -106,11 +106,11 @@ class BountyService {
     );
   }
 
-  Future<String> approveBounty(int submissionId) {
+  Future<String> approveBounty(BigInt submissionId) {
     return _clientService.approveBounty(submissionId);
   }
 
-  Future<String> contibuteToBounty(int bountyId, int amount) {
+  Future<String> contibuteToBounty(BigInt bountyId, BigInt amount) {
     return _clientService.contibuteToBounty(bountyId, amount);
   }
 }
